% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph-style.R
\name{igraph_style}
\alias{igraph_style}
\title{The igraph style guide}
\usage{
igraph_style(scope = "tokens", ...)
}
\description{
The igraph style guide
}
\examples{
styler::style_text("A+-+B", style = igraph.style::igraph_style)
}
